% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_analyticity.R
\name{ck_analyticity}
\alias{ck_analyticity}
\title{Analyticity checking function for the Cauchy-Kowalevski theorem
Theorem : A function f is analytic at a point x₀ if it can be represented by a power series that converges in a neighborhood of that point.
Algorithm : ck_analyticity returns true verify continuity, singularity, derivability and taylor are true.
NB : Equation should be of the form F(x, y, u, u_x, u_y, u_xx, u_xy, u_yy) = 0}
\usage{
ck_analyticity(
  f,
  point,
  variables = c("x", "y", "u", "ux", "uy", "uxx", "uxy", "uyy"),
  epsilon = 0.001,
  ordre_taylor = 5
)
}
\arguments{
\item{f}{Function to be checked (R function or an expression)}

\item{point}{Point around which to check analyticity (named vector)}

\item{variables}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}

\item{ordre_taylor}{Order of Taylor series development to be verified}
}
\value{
List with verified results
}
\description{
Analyticity checking function for the Cauchy-Kowalevski theorem
Theorem : A function f is analytic at a point x₀ if it can be represented by a power series that converges in a neighborhood of that point.
Algorithm : ck_analyticity returns true verify continuity, singularity, derivability and taylor are true.
NB : Equation should be of the form F(x, y, u, u_x, u_y, u_xx, u_xy, u_yy) = 0
}
