% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_conditions.R
\name{ck_conditions}
\alias{ck_conditions}
\title{Complete verification of the 3 conditions and theorem application}
\usage{
ck_conditions(
  edp_fonction,
  edp_coefficients,
  point_test_edp,
  variables_edp,
  phi,
  psi = NULL,
  points_test_CI,
  variables_spatiales,
  variable_temps = "t",
  epsilon = 0.001,
  ordre_taylor = 3
)
}
\arguments{
\item{edp_fonction}{The PDE fonction to be check}

\item{edp_coefficients}{PDEs coefficients to be verified (R function or an expression or constant)}

\item{point_test_edp}{Point around which to check CK conditions}

\item{variables_edp}{Variables of the PDE Equation}

\item{phi}{Function to be checked (R function or an expression)}

\item{psi}{du/dxi Function to be checked if PDE is of order 2 (R function or an expression)}

\item{points_test_CI}{Point around which to check analyticity (named vector)}

\item{variables_spatiales}{Name of variables (default for dim 2)}

\item{variable_temps}{time variable}

\item{epsilon}{Neighborhood radius to test}

\item{ordre_taylor}{Order of Taylor Series to be verified}
}
\value{
List with verified results
}
\description{
Complete verification of the 3 conditions and theorem application
}
