% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_analyticity.R
\name{verify_analyticity_CK}
\alias{verify_analyticity_CK}
\title{Solve Cauchy-Kovalevskaya theorem equation
Condition 1:
Analyticity checking function for the Cauchy-Kowalevski theorem
Order 2: Equation of the form F(x, y, u, u_x, u_y, u_xx, u_xy, u_yy) = 0}
\usage{
verify_analyticity_CK(
  f,
  point,
  variables = c("x", "y", "u", "ux", "uy", "uxx", "uxy", "uyy"),
  epsilon = 0.001,
  ordre_taylor = 5
)
}
\arguments{
\item{f}{Function to be checked (R function or an expression)}

\item{point}{Point around which to check analyticity (named vector)}

\item{variables}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}

\item{ordre_taylor}{Order of Taylor series development to be verified}
}
\value{
List with verified results
}
\description{
Solve Cauchy-Kovalevskaya theorem equation
Condition 1:
Analyticity checking function for the Cauchy-Kowalevski theorem
Order 2: Equation of the form F(x, y, u, u_x, u_y, u_xx, u_xy, u_yy) = 0
}
