% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_continuity.R
\name{verify_continuity}
\alias{verify_continuity}
\title{Continuity of f}
\usage{
verify_continuity(f, point, variables, epsilon)
}
\arguments{
\item{f}{Function to be checked (R function or an expression)}

\item{point}{Point around which to check analyticity (named vector)}

\item{variables}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}
}
\value{
List with verified results
}
\description{
Continuity of f
}
