% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_derivability.R
\name{verify_derivability}
\alias{verify_derivability}
\title{Differentiability of f}
\usage{
verify_derivability(f, point, variables, h = 1e-04)
}
\arguments{
\item{f}{Function to be checked (R function or an expression)}

\item{point}{Point around which to check analyticity (named vector)}

\item{variables}{Name of variables (default for dim 2)}

\item{h}{step size for differentiation}
}
\value{
List with verified results
}
\description{
Differentiability of f
}
