% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_conditions.R
\name{verify_initial_conditions_CK}
\alias{verify_initial_conditions_CK}
\title{Initial conditions Phi and Psi are analytic}
\usage{
verify_initial_conditions_CK(
  phi,
  psi = NULL,
  points_test,
  variables_spatiales = c("x"),
  epsilon = 0.001,
  ordre_taylor = 3
)
}
\arguments{
\item{phi}{Function to be checked (R function or an expression)}

\item{psi}{du/dxi Function to be checked if PDE is of order 2 (R function or an expression)}

\item{points_test}{Point around which to check analyticity (named vector)}

\item{variables_spatiales}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}

\item{ordre_taylor}{Order of Taylor Series to be verified}
}
\value{
List with verified results
}
\description{
Initial conditions Phi and Psi are analytic
}
