% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_surface.R
\name{verify_non_characteristic_surface}
\alias{verify_non_characteristic_surface}
\title{Condition 3: Non-characteristic surface}
\usage{
verify_non_characteristic_surface(
  edp_coefficients,
  point_initial,
  variable_temps = "t",
  variables_spatiales = c("x", "y"),
  epsilon = 0.001
)
}
\arguments{
\item{edp_coefficients}{PDEs coefficients to be verified (R function or an expression or constant)}

\item{point_initial}{Point around which to check analyticity (named vector)}

\item{variable_temps}{time variable}

\item{variables_spatiales}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}
}
\value{
List with verified results
}
\description{
Condition 3: Non-characteristic surface
}
