% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_taylor.R
\name{verify_taylor}
\alias{verify_taylor}
\title{Taylor series converges to f}
\usage{
verify_taylor(f, point, variables, epsilon, ordre)
}
\arguments{
\item{f}{Function to be checked (R function or an expression)}

\item{point}{Point around which to check analyticity (named vector)}

\item{variables}{Name of variables (default for dim 2)}

\item{epsilon}{Neighborhood radius to test}

\item{ordre}{Order of Taylor Series to be verified}
}
\value{
List with verified results
}
\description{
Taylor series converges to f
}
